requirements.txt
flask==3.0.3
python-dotenv==1.0.1
groq==0.12.0
requests==2.32.3
gunicorn==23.0.0
sendgrid==6.11.0
